integrator: without filtering
Since the integrator is a huge filter (accumulator), there is no advantage to filter the input signal of it.

To use this type filtering may improves control stability when applied on the proportional part. But this will delay the response time by 1/2 (?) control step (may you don't care in your use case).