Fix issue with angle wrap-around
The 0.8 factor was sub-optimal.

Let's assume an angular velocity of 0.3*2PI per calls of update.
Then a wrap-around will produce a change of 2PI - 0.3*2PI, that's
0.7*2PI
So it won't be detected as a wrap-around of the angle at all and
thus the number of full rotations will be off, as will be the
velocity.

The perfect number is 0.5.